requirements.txt
asgiref==3.5.2
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
chardet==2.1.1
chardet2==2.0.3
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
Faker==13.13.0
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
protobuf==3.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-twitter==3.5
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
